+++ Maya/Maya2Babylon2017.csproj
@@ -83,6 +83,8 @@
     <Compile Include="Exporter\BabylonExporter.Material.Duplication.cs" />
     <Compile Include="Exporter\MaterialDuplicationData.cs" />
     <Compile Include="Exporter\PluginNodes\babylonAiStandardSurfaceMaterialNode.cs" />
+    <Compile Include="Exporter\PluginNodes\babylonMaterialNodeBase.cs" />
+    <Compile Include="Exporter\PluginNodes\babylonMPxNode .cs" />
     <Compile Include="Exporter\PluginNodes\babylonStandardMaterialNode.cs" />
     <Compile Include="Exporter\PluginNodes\babylonStingrayPBSMaterialNode.cs" />
     <Compile Include="Extensions\ColorExtension.cs" />
